requirements.txt
aiofiles==23.2.1
aioredis==2.0.1
async-timeout==4.0.3
Authlib==1.3.1
blinker==1.8.2
cachetools==5.3.3
certifi==2024.6.2
cffi==1.16.0
charset-normalizer==3.3.2
click==8.1.7
cryptography==42.0.7
exceptiongroup==1.2.1
Flask==3.0.3
google-auth==2.29.0
google-auth-oauthlib==1.2.0
h11==0.14.0
h2==4.1.0
hpack==4.0.0
Hypercorn==0.17.3
hyperframe==6.0.1
idna==3.7
itsdangerous==2.2.0
Jinja2==3.1.4
looseversion==1.3.0
MarkupSafe==2.1.5
oauthlib==3.2.2
passlib==1.7.4
priority==2.0.0
pyaml==24.4.0
pyasn1==0.6.0
pyasn1_modules==0.4.0
pycparser==2.22
pyhumps==3.8.0
PyJWT==2.8.0
python-dotenv==1.0.1
PyYAML==6.0.1
Quart==0.19.6
quart-schema==0.20.0
requests==2.32.3
requests-oauthlib==2.0.0
rethinkdb==2.4.10.post1
rsa==4.9
six==1.16.0
taskgroup==0.0.0a4
tomli==2.0.1
typing_extensions==4.12.1
urllib3==2.2.1
Werkzeug==3.0.3
wsproto==1.2.0
quart-schema[pydantic]
